feat: Add Git tutorial page
The code changes include adding a new Git tutorial page to the tutorials section. This page is represented by the `git.mdx` file and is accompanied by a `_meta.json` file that provides metadata for the page.

Recent user commits and repository commits do not provide any relevant information for generating the commit message.